Fix includes related to metal view
