+++ shards/gfx/platform_surface.hpp
@@ -16,22 +16,6 @@
 #endif
 
 namespace gfx {
-
-#if SH_APPLE
-struct MetalViewContainer {
-  SDL_MetalView view{};
-  void *layer{};
-
-  MetalViewContainer(SDL_Window *window) {
-    view = SDL_Metal_CreateView(window);
-    layer = SDL_Metal_GetLayer(view);
-  }
-  ~MetalViewContainer() { SDL_Metal_DestroyView(view); }
-  MetalViewContainer(const MetalViewContainer &other) = delete;
-  operator void *() const { return layer; }
-};
-#endif
-
 struct WGPUPlatformSurfaceDescriptor : public WGPUSurfaceDescriptor {
   union {
     WGPUChainedStruct chain;
